refactor: home sections
