--- packages/gdsc-dju-web/src/components/HomeSection/SectionTimeLine.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SectionTimeLine = () => {
-  return <div>timeLine</div>;
-};
-
-export default SectionTimeLine;
